refactor(mcp): split order resource definition
- move the dynamic order MCP resource into order-resource
- keep resources.ts focused on static menu and open-order resources
- preserve the public OrderResource export through resources.ts
